WMP001-015 Transaction amount na double setter ograniczyć do czesci s…

+++ src/main/java/com/kodilla/walletmanager/dto/TransactionDto.java
@@ -11,5 +11,5 @@ public class TransactionDto {
     private Long id;
     private String title;
     private String description;
-    private long amount;
+    private double amount;
 }
